MDL-60978 phpunit: Enable output buffering
This is a workaround for MDL-60951, to prevent output from being flushed,
which would signal headers_sent and cause calls to session_set_save_handler
to trigger a warning and fail. PHP 7.2 wants to make sure session functions
don't pretend to succeed when they have actually failed, but for CLI operations
headers are never returned as output so it points to a bug (php/php-src#75628).