MDL-66107 js: Stop bundling sourceContent in JS Maps
The commit removes the `sourceContent` from the generated map files
during run-time as this is no longer required.

When we first landed our Babel transpilation for Moodle, we attempted to
include combined map files, whereby the combined first.js file had a
single sourceMap referencing all of the source files.

This worked on Chrome, but did not work in Firefox due to a bug with
their sourceMap support.

See https://bugzilla.mozilla.org/show_bug.cgi?id=1556903 for more
information that bug.

As a compromise we only put sourceMaps into development code, and we
included the sourceContent within the generated map. This allowed us to
work around that bug entirely and improve the development experience in
Firefox.

That issue how now been resolved for some time, and all affected
versions of Firefox are long out of use and support. This means that we
no longer need to include the sourceContent in the generated sourcemap.

Although it is possible to remove this in the build process too, doing
so will lead to changes in build output which will affect developers. We
should avoid making this change until we need to make other changes to
the build output and make these changes all at once.

By removing the sourceContent from the generated sourcemap at runtime,
we can at least reduce over-the-wire size, which is still beneficial to
developers.