MDL-66107 js: Turn off productionsourcemaps by default
This commit removes sourcemaps from production by default.

Production and development builds are almost identical. The only difference
between them previously was the inclusion of the source map, which was
previously excluded from production builds due to concerns about the size
of the sourcemap, and inclusion of it in production.

I feel that these concerns are no longer relevant, and there is no real
negative to having source maps in our builds:
- They are only loaded lazily when opening the relevant files in the
  developer tooling;
- they no longer include the sourceContent in them, so they are
  relatively small;
- they assist in debugging issues in production code on production sites
  with low risk; and
- source maps do not expose any additional content that is not already
  available so there is no security risk to this change either.

Source maps can now be configured to be used in production code, to
enable easier production debugging.

This change has the benefit that developers can now work with, and test
production code, rather than code which may differ in production vs a
development environment.