MDL-66107 js: Always lazyload on modern HTTP protocols
This commit prioritizes lazyload for modern HTTP protocols, avoiding
combo-loading of all JS code into a single file (first.js). Instead,
modules will be fetched individually.

Upon further investigation, it seems that the use of a combo-loader is a
false economy in most cases. As JS usage has increased, we have more and
more code being needlessly downloaded for content which is not used by
the end-user. Most pages use the same set of common modules, plus code
specific to that page. These common modules are cached by the browser
upon first opening of the page, and only additional missing modules are
fetched. Usually these are relatively trivial in nature.

MDL-66107 js: Remove support for legacy AMD modules (pre-3.8)

This commit removes support for legacy AMD modules combined with Moodle
versions older than 3.8.

In Moodle 3.8, with MDL-62497, we rewrote the module compilation process
and moved from serving source files from the **/amd/src/ directory, to
serving minified files served from the **/amd/build/ directory which has
been transpiled by Grunt + babel.

As an interim solution to assist developers in the migration, we
supported adding the module name to the build. This was intended to
support developers compiling their code with older versions of Moodle to
aid in transition. This was intended specifically for Moodle versions
from 3.5 (LTS) to 3.7 which were the supported Moodle versions at the
time of the 3.8 release.

Moodle 3.8, where this change was originally introduced, was first
released nearly three years ago, and support ended 18 months ago. Moodle
3.5 (LTS) support also ended at the same time.

Now that these older versions of Moodle are long out of support, I feel
that it is reasonable to drop support of this transitional assistance.

It is worth noting that we further improved compatibility with
third-party JS code shipped as an AMD module in MDL-71882, which affects
Moodle 3.9 and up. This further cements the removal of the legacy
support code as it performs the job of that code of replacing any
incorrect module name.

Therefore I feel it is entirely safe, and reasonable, to remove support
for these legacy AMD modules, where no module name was specified.